refactor(validations): moving validations to src
