+++ src/helpers/index.js
@@ -1,11 +1,9 @@
 const { queryHelper } = require('./queryHelper');
 const { encryptor } = require('./encryptor');
-const { validationSchemas } = require('./validations');
 const { mailer } = require('./mailer');
 
 module.exports = {
   queryHelper,
   encryptor,
-  validationSchemas,
   mailer,
 };
